Check for `AxProtector_*` files instead of `.wbc`
Some projects don't contain a `.wbc` file in the required format. But they
usually contain a `AxProtector_<project>.xml` which acts as a requirement now
in MORYX 6 projects for determining if a project should be protected by a
license.

If it doesn't exist, the licensing procedure would exit with an error code, in
order to fail the pipeline.